Modify the attack coefficient from a float to an int type

+++ Assets/2.Script/Data/AttackStat.cs
@@ -9,7 +9,7 @@ namespace Data
         public string Name;
         public string AnimationKey;
         public float Cooldown;
-        public Dictionary<int, float> CoeffDictionary;
+        public Dictionary<int, int> CoeffDictionary;
         public EAttackType AttackType;
         public int Range;
         public int ProjectileID;
